server: Add Validate and Apply for policy
Validate allows a user to build an example AttestationPolicy
and determine whether it would be valid against a real
machine (e.g., checks that the db is not empty).

Apply takes an AttestationPolicy and MachineState and returns
a list of errors on applying the policy to the MachineState.

This commit also contains DefaultGceLinuxPolicy, which is
a partial GCE Linux VM Policy, without distro-specific certs
in the db (used by shim).